# backend/docs/api/openapi/openapi.yaml
...
{
  "openapi": "3.1.0",
  "info": {
    "title": "ManifestPay API",
    "description": "AI-Powered Payment Infrastructure for Autonomous Agents on Stellar",
    "version": "1.0.0",
    "contact": {
      "name": "ManifestPay Team",
      "url": "https://github.com/frienzy514-png/manifestpay",
      "email": "hello@manifestpay.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.manifestpay.com/api/v1",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/sandbox/status": {
      "get": {
        "tags": [
          "Sandbox"
        ],
        "summary": "Sandbox environment status",
        "operationId": "GET_/sandbox/status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [],
        "deprecated": false
      }
    },
    "/sandbox/payments/process": {
      "post": {
        "tags": [
          "Sandbox",
          "Payments"
        ],
        "summary": "Process mock payment (sandbox)",
        "operationId": "POST_/sandbox/payments/process",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [],
        "deprecated": false
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key authentication"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "VALIDATION_FAILED"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "statusCode": {
                  "type": "number",
                  "example": 404
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized access",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "statusCode": {
                  "type": "number",
                  "example": 401
                }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "statusCode": {
                  "type": "number",
                  "example": 500
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}